[Geo] Add elementary geometry structures and functions
Add Point, Segment, and Line structs. Focusing on testing Point and
Segment structs for now. The algorithms of interest are more related to
segments. We can revisit Line at another time.

Lots of edge cases are not handled well. For example, it is difficult to
handle verticle lines because the slope is infinite. We will come back
to these.