[Hash] Refactor to use linked lists and provide equality comparator
RBTree was awkward because it required specifying an Ordering
comparator. We could use the hash of each as an Ordering but the problem
is we would have to generate a lambda that takes in the hash function as
a reference, which C does not have. We also cannot pass in the equality
comparator for the same reasons.